test(reward): move epoch reward restart test to slow-tests
- relocate from espresso-node lib tests to slow-tests integration crate
- runs under the `slow` nextest profile (no per-test kill, 90m job cap)
- avoids the 6m default slow-timeout that terminated it in sharded CI
- rename to slow_test_epoch_reward_restart per slow-tests convention
